requirements-backend.txt
fastapi>=0.100.0,<0.101.0
uvicorn>=0.22.0,<0.23.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.11.0,<1.12.0
psycopg2>=2.9.6,<2.10.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.6,<0.0.7
requests>=2.31.0,<2.32.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<1.1.0
asyncpg>=0.27.0,<0.28.0
redis>=4.6.0,<4.7.0
httpx>=0.24.1,<0.25.0
tenacity>=8.2.3,<8.3.0